New arrangement of the code, using URL Rules and Class-Based Views
